Updated localisation for new bait subclass

+++ Content/Items/Weapons/Bait/AntlionShell.cs
@@ -17,16 +17,22 @@ namespace CalamityEntropy.Content.Items.Weapons.Bait
 {
     public class AntlionShell : ModItem, IBaitItem
     {
-        public static int TagDamage = 4;
+        public static int TagDamage = 2;
+        public static float DamageMult = 2;
         public override LocalizedText Tooltip => base.Tooltip.WithFormatArgs(TagDamage);
 
         public override void SetDefaults()
         {
+            Item.damage = 20;
+            Item.knockBack = 0;
+            Item.shootSpeed = 26;
+            Item.useAnimation = Item.useTime = 20;
+            Item.value = CalamityGlobalItem.RarityGreenBuyPrice;
+            Item.rare = ItemRarityID.Green;
             Item.width = 38;
             Item.height = 38; 
             Item.autoReuse = false;
             Item.useStyle = ItemUseStyleID.Swing;
-            Item.useAnimation = Item.useTime = 20;
             var snd = CEUtils.GetSound("BaitThrow", 1, 6);
             snd.PitchRange = (0f, 0.4f);
             Item.UseSound = snd;
@@ -34,20 +40,15 @@ namespace CalamityEntropy.Content.Items.Weapons.Bait
             Item.DamageType = DamageClass.SummonMeleeSpeed;
             Item.noUseGraphic = true;
             Item.shoot = ModContent.ProjectileType<AntlionShellProjectile>();
-            Item.damage = 10;
-            Item.knockBack = 0;
-            Item.shootSpeed = 24;
-            Item.value = CalamityGlobalItem.RarityGreenBuyPrice;
-            Item.rare = ItemRarityID.Green;
             Item.autoReuse = true;
         }
         public override bool CanUseItem(Player player)
         {
-            return player.Entropy().BaitCharge >= 1;
+            return player.Entropy().BaitUsable;
         }
         public override bool Shoot(Player player, EntitySource_ItemUse_WithAmmo source, Vector2 position, Vector2 velocity, int type, int damage, float knockback)
         {
-            Projectile.NewProjectile(source, position, velocity, type, damage, knockback, player.whoAmI, 0, 0, TagDamage);
+            Projectile.NewProjectile(source, position, velocity, type, (int)(damage * DamageMult), knockback, player.whoAmI, 0, 0, TagDamage);
             return false;
         }
 
@@ -94,10 +95,10 @@ namespace CalamityEntropy.Content.Items.Weapons.Bait
                     return;
                 }
                 Main.player[Projectile.owner].MinionAttackTargetNPC = npc.whoAmI;
+                npc.GetGlobalNPC<WhipDebuffNPC>().BaitStick = 2;
                 if (IsActive)
                 {
                     Projectile.GetOwner().Calamity().mouseWorldListener = true;
-                    npc.GetGlobalNPC<WhipDebuffNPC>().BaitStick = 3;
                     npc.GetGlobalNPC<WhipDebuffNPC>().ClearBaitTags();
                     npc.GetGlobalNPC<WhipDebuffNPC>().Tags.Add(new WhipTag(this.GetType().Name, 5, this.TagDamage, 1, 0, this.GetType().Name) { IsABaitTag = true });
                 }
@@ -107,26 +108,24 @@ namespace CalamityEntropy.Content.Items.Weapons.Bait
                 {
                     if(IsActive)
                     {
-                        IsActive = false;
                         CEUtils.SyncProj(Projectile.whoAmI);
-                        ActiveEffect();
+                        SetActive();
                     }
                 }
             }
             activeEffectAlpha = float.Lerp(activeEffectAlpha, (StickNPC >= 0 && IsActive) ? 1 : 0, 0.04f);
             Counter++;
         }
-        public void ActiveEffect()
+        public override void ActiveEffect(float damageMul)
         {
             if(Main.myPlayer == Projectile.owner)
             {
-                Projectile.NewProjectile(Projectile.GetSource_FromAI(), Projectile.Center + new Vector2(Main.rand.NextFloat(-400, 400), 620), Vector2.Zero, ModContent.ProjectileType<DesertNuisanceFriendly>(), Projectile.damage * 9, 6, Projectile.owner);
+                Projectile.NewProjectile(Projectile.GetSource_FromAI(), Projectile.Center + new Vector2(Main.rand.NextFloat(-400, 400), 700), Vector2.Zero, ModContent.ProjectileType<DesertNuisanceFriendly>(), (int)(Projectile.damage * damageMul), 6, Projectile.owner);
             }
         }
         public float activeEffectAlpha = 0;
         public override bool PreDraw(ref Color lightColor)
         {
-            Main.EntitySpriteDraw(Projectile.getDrawData(lightColor));
             if (activeEffectAlpha >= 0.01f)
             {
                 Main.spriteBatch.UseAdditiveClamp();
@@ -138,6 +137,7 @@ namespace CalamityEntropy.Content.Items.Weapons.Bait
                 }
                 Main.spriteBatch.ExitShaderRegion();
             }
+            Main.EntitySpriteDraw(Projectile.getDrawData(lightColor));
             return false;
         }
         public override bool? CanDamage()
@@ -146,11 +146,12 @@ namespace CalamityEntropy.Content.Items.Weapons.Bait
         }
         public override void OnHitNPC(NPC target, NPC.HitInfo hit, int damageDone)
         {
+            Projectile.tileCollide = false;
             OnHitEffect(Projectile.Center);
             Projectile.velocity *= 0;
             StickNPC = target.whoAmI;
             StickOffset = Projectile.Center - target.Center;
-            Projectile.timeLeft = 360;
+            Projectile.timeLeft = 480;
             CEUtils.SyncProj(Projectile.whoAmI);
         }
         public void OnHitEffect(Vector2 pos)
@@ -171,7 +172,7 @@ namespace CalamityEntropy.Content.Items.Weapons.Bait
         }
         public override void OnKill(int timeLeft)
         {
-            if(timeLeft > 0)
+            if(timeLeft > 0 && !Main.dedServ)
             {
                 OnHitEffect(Projectile.Center);
             }
@@ -195,13 +196,15 @@ namespace CalamityEntropy.Content.Items.Weapons.Bait
             Projectile.FriendlySetDefaults(DamageClass.Summon, false, -1);
             Projectile.width = Projectile.height = 64;
             Projectile.localNPCHitCooldown = -1;
-            Projectile.timeLeft = 200;
+            Projectile.timeLeft = 260;
         }
         public int headFrame = 0;
         public bool Bite = true;
         public bool InGround = true;
         public override void AI()
         {
+            if (Projectile.timeLeft < 18)
+                Projectile.Opacity -= 1 / 18f;
             if (spawnSeg)
             {
                 spawnSeg = false;
@@ -227,7 +230,7 @@ namespace CalamityEntropy.Content.Items.Weapons.Bait
                 {
                     targetPos = player.Calamity().mouseWorld;
                 }
-                Projectile.velocity = (targetPos - Projectile.Center).normalize() * 30;
+                Projectile.velocity = (targetPos - Projectile.Center).normalize() * 36;
             }
 
             if (player.MinionAttackTargetNPC >= 0 && Bite)
@@ -256,7 +259,7 @@ namespace CalamityEntropy.Content.Items.Weapons.Bait
                     {
                         if(p.ModProjectile != null && p.ModProjectile is BaitProj ibp)
                         {
-                            if (!ibp.IsActive)
+                            if (!ibp.IsActive && ibp.StickNPC == player.MinionAttackTargetNPC)
                                 p.Kill();
                         }
                     }
@@ -345,10 +348,9 @@ namespace CalamityEntropy.Content.Items.Weapons.Bait
 
             return false;
         }
-        public Texture2D texGlow => CEUtils.getExtraTex("NxDragonGlow");
         public void DrawSeg(Texture2D tex, Vector2 pos, Rectangle? frame, float rot, Vector2 origin, Color color)
         {
-            Main.EntitySpriteDraw(tex, pos - Main.screenPosition, frame, color, rot + MathHelper.PiOver2, origin, Projectile.scale, SpriteEffects.None);
+            Main.EntitySpriteDraw(tex, pos - Main.screenPosition, frame, color * Projectile.Opacity, rot + MathHelper.PiOver2, origin, Projectile.scale, SpriteEffects.None);
         }
     }
 }
